Additional reporting (#526)
* Add additional penalty vector to Result object

additional_penalty is a vector with the number of element equal to the number of penalties in the model scheme

* Fixed broken docs

Some documentation was outdated

Tested locally using
python -m sphinx -W -b html docs/source docs/build/html
after installing MiKTeX for the latex math images.

* Fixed error in printing additional penalties when None

In e.g. the quickstart additional penalties are not used so set to None thus should not be attempted to print

* Added weighted_root_mean_square_error to Result

Added weighted_root_mean_square_error calculation to _create_result_dataset In problem class which is added to the Result object in the end

The new attribute is used in Markdown printing (e.g. to .md file)