Use a notebook for quickstart (#568)
Our Quickstart documentation is now an executable document (.ipynb notebook)


Detailed changelog:
* Changed pre-commit config in preparation for jupyter notebooks
* Added jupyterlab3 as dev requirement for docs
* Basic rewrite of quickstart as a notebook
* Added nc files generated by notebooks to .gitignore
* Added nbsphinx and other notebook doc dependencies aswell as configs
* Added headings for quicker navigation
* Added some jupyter helpers to make the output prettier
And made flake8 for notebooks ignore 'E402 module level import not at top of file'
* Added binder config
* Replaced 'IPython.display.Code' with custom 'print_yaml_file' function
The problem with 'Code' was, while it was properly syntax highlighted in the Html docs, it doesn't work in jupyterlab (see: issue 11747 in ipython/ipython) and the workaround is longer than the markdown wrapping in 'print_yaml_file'. Also, the colormap used by 'pygments' does not adjust or work well with the dark theme.
* Added pandoc to doc related CI tasks, so the docs will actually build
* Added tests for the notebooks, that way errors can be easier spotted
* Adjusted notebook to API changes
* Added pre-commit hook to remove empty notebook cells
* Use tox-direct when no seperate env is needed
E.g. docs, docs-links, docs-notebooks, pre-commit since those should all be runnable in the local devenv.
The main reason to use tox-direct is that building the docs fails with image path resolving errors for the notebooks.
* Added Trigger-Binder-build workflow and binder badge
That way users don't need to wait for the binder image to build, since it gets built each push to main
* Added ast-check and isort to notebook pre-commit hooks
* Removed usage warning
* Replaced my feature branch deprendency of tox-direct with new release on PyPi