♻️Moved exception texts to io interface classes
and added default docstrings, so the docs indicate if a method was implemented ore not.

Note:
For read methods '.. # noqa: DAR202' is needed since the template implementation doesn't return anything and '.. # noqa: DAR401' since the 'NotImplementedError' isn't mentioned in the docstring.
For write methods '.. # noqa: DAR401' is needed since the 'NotImplementedError' isn't mentioned in the docstring and  '.. # noqa: DAR101' is needed due to a bug in darglint
See issue 69 at https://github.com/terrencepreilly/darglint