✨Added protect_from_overwrite fuction to all Io write functions
The protect_from_overwrite from overwrite fuction will raise FileExistsError if the passed path points to an existing file or an existing folder which isn't empty.
Since this is implemented in the convenience level functions and not the plugin itselfe there will be consisten behaviour across plugins.
Even if a plugin developer forgets to protect the user from accidently overwriting existing files, we do!