♻️ Refactor Result and Scheme loading to to use 'file' fields (#903)
* ♻️👌 Removed file fields in ProjectIo like classes and used unified field

If the field isn't an instance of the targetClass it will try to load the class instance from file. This also allows classes like scheme to be initialized with file paths directly.
When dataclasses with file loadable fileds are serialized the objects will be replaced with their source path.

In addition projectIO load and save functions will set the ``source_path`` attribute of the class instances.

* ♻️🔌 Refactored load_dataset to always return xr.Dataset

* ♻️ Added type 'StrOrPath' and refactored io plugins with new type

Also ensure that paths are passed as posix formatted path

* ✨ Implemented convenience function 'load_datasets'

and file loadable wrapper class 'DatasetMapping'.

This also allows to load all datasets need for a opimization directly from file path passing a dict with the keys used for the dataset names and the paths as values.

* ♻️👌 Made 'DatasetMapping.source_path' a property accessing the dataset

This way it will update when the source_path of the dataset is updated e.g. by calling 'save_dataset'.

* ♻️👌 Replaced all file_representation_field  with file_loadable_field

Also, removed file_representation_field compleatly.

* ♻️✨ Factored making paths relative and posix style out

and added support  for Sequence like FileLoadable classes

* ♻️ Refactored bool_str_repr after sourcery suggested a differet change

* ♻️🩹 Changed implementation of relative_posix_path to use os.path.relpath

instead of pathlib.Pathrelative_to This prevents crashes as long at the files are on the same drive.